admin/requirements.txt
asgiref==3.6.0
Django==3.2
django-postgres-extra==2.0.8
django-split-settings==1.2.0
flake8==4.0.0
mccabe==0.6.1
psycopg2-binary==2.9.9
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
gunicorn==19.8.1
requests==2.31.0